Add illink analyzer support for field/property initializers (#92600)
The dataflow analyzer was exiting early for cases where the owning
symbol was not an `IMethodSymbol`. This meant we weren't running
dataflow analysis for field and property initializers.

This fixes it by allowing through cases where the owning symbol is not
an `IMethodSymbol`, and adding testcases to validate that we don't hit
asserts in the code paths that only light up for methods.