Fix compilation failure in the 22888 regression test
This failure was caused by incorrectly making a test out of what
is in reality a library used by custom means by the test app
(for the purpose of accessing resources). While I believe we should
further clean this logic up, this should be good enough to make
tests pass.

Thanks

Tomas