Get rid of static constructor in Task<T>
Because the `default(TResult)` task was cached in in the
`Task<TResult>` class, every instantiation would run the static
constructor.

With this change, the cached value is stored in
`TaskCache.CachedDefaultValueTask<TResult>` instead, and
`Task<TResult>` no longer needs to execute a static constructor.

This is expected to improve startup time for applications that do a
lot of async work.  (But no measurements have been done)