Remove dependencies on non-live illink (#91468)
Removes remaining dependencies on non-live ILLink.

The live ILLink logic is consolidated into liveILLink.targets,
which gets imported from source projects that depend on ILLink.

This removes the dependency on UsingToolMicrosoftNetILLinkTasks,
and turns off ILLink dependency flow.

* Don't require illink analyzers for source generator projects

Microsoft.Extensions.Options.SourceGeneration was enabling
the single-file analyzer, but did not import illink.targets
because it is not a source project.

We could fix this by importing liveILLink.targets instead, but
this introduces more problems with circular dependencies
(generator projects depending on illink depending on
generator projects).

* Only enable single-file analyzer for source projects

This more closely matches the condition on the import of
illink.targets. When EnableSingleFileAnalyzer is true, we should
also import liveILLink.targets so that we get the live analyzer
bits.

* Remove import from tests.mobile.targets

Mobile trimmingtests should already use the live illink with
the logic in project.csproj.template. The import of illink.targets
shouldn't be required, unless tests.mobile.targets are required
for other test scenarios.

This import was resulting in a double-import, and wasn't working
because liveILLink needs ToolsILLinkDir to be set, but it's not
set during the trimming test build.

---------

Co-authored-by: Viktor Hofer <viktor.hofer@microsoft.com>