JIT: Fix emitOutputLJ trying to encode large displacements
For hot<->cold jumps, emitOutputLJ would try to encode the displacement
as part of the jump instruction. This is not generally possible since
the cold and hot code buffers can be allocated at arbitrary addresses by
the EE. We expect the EE to handle the displacement as part of
relocations, so simply encode 0 for the cross jumps.

Also remove the ifndef below; the comment on it is wrong since we
usually do not call recordRelocation there. We're currently always
recording relocations when cross jumps are possible (this might change
if we do JIT time hot cold splitting).

Fix #93166