JIT: defer creating throw helper blocks until simple lower
Don't create any throw helper blocks until simple lowering. This removes
some of the flow graph updates from morph, with the am of making it easier
to show that an RPO traversal in morph is safe.

Unfortunately it's hard to get these blocks in the exact same places
they were before, so there are some diffs.