[wasm] Runtime tests - disable webcil
The runtime tests build for wasm have been timing out on CI. It seems that the
webcil conversion part is causing the memory limits on the container to
be hit. Disable this temporarily to get the runtime tests building
again.

In a follow up PR, this part can be moved to helix, same as we do for
non-merged test runners.

Issue: https://github.com/dotnet/runtime/issues/93134