Enable cgroups hierarchical limits support
Cgroups can limit memory hierarchically. That means that ancestor
cgroups can apply limits that are lower than the current cgroup's
memory limit. .NET runtime is not honoring these hierarchical limits
with neither cgroups v1 nor cgroups v2.
This change fixes it. The implementation is different for cgroups V1
and V2 because cgroups v1 provide aggregated hierarchical limit value
in the memory.stat, but in cgroups v2, runtime needs to walk the
cgroup hierarchy and find the lowest limit in it.