Update test build group boundaries to today test distribution
Test build group boundaries are several years old; since then, the
JIT/HardwareIntrinsics tests have arrived and substantially changed the
test distribution. Moreover, as by now the vast majority (over 95%)
of tests use the new-style merged wrappers, it makes no sense to
establish build group boundaries in the middle of merged wrappers
because that only means that some test projects need to be traversed
twice, once in the group they belong to according to the partitioning
and one more time as a dependency of their merged test wrapper.

Thanks

Tomas