Only write cmake_cmd_line.txt cache file if cmake succeeded
I noticed that we wrote the cache file before invoking cmake, this meant that if the cmake configure step failed (e.g. because of an invalid command like in #93670) then running the build again would skip cmake configure, resulting in a broken build setup.

This only applies to non-VS generators like ninja.