JIT: fix bug introduced by recent throw helper block changes
`compUsesThrowHelper` indicates codegen will likely need to create a call to a helper,
whether or not such calls are shared.

In #93371 I didn't appreciate this, and so the JIT was failing to set `compUsesThrowHelper`
in cases where a throw helper was going to be needed but throw helper calls were not going to
be shared.

Fixes #93710.