Suggest using the latest C # syntax
In C # 9.0 and later versions, a syntax sugar called "target typed new expression" was introduced, allowing you to omit the parameter list of the constructor when instantiating an object. Simply use the parentheses () after the new keyword, and the compiler will infer the type of object to be instantiated based on the variable or class on the left. The main purpose of this syntax sugar is to make the code more concise and readable.