Fix analyzer assert and add dynamic object tests (#94058)
Fixes the assert from
https://github.com/dotnet/runtime/issues/94016. https://github.com/dotnet/runtime/issues/94057
tracks adding full support for this in the analyzer.

We had some existing tests that only covered
IDynamicInvocationOperation. Those have been moved to a shared
testcase (shared by ILLink/NativeAot), and some additional
testcases have been added that include coverage for
IDynamicMemberReference, which was causing the assert.

This includes the same fix for IDynamicIndexerAccessOperation.