Fix the type of handle constants replacing TYP_REF expressions
`optVNConstantPropOnTree` will propagate a constant handle value
into a tree node currently typed as TYP_REF, where the new constant
node is TYP_I_IMPL. This leads to trees like:

```
*  STORE_LCL_VAR ref   (AX) V08 tmp6
\--*  CNS_INT(h) long   $181
```

This seems like an undesired type mismatch. In another change, code
was asserting because of this.

This change causes diffs: the constant is GC reported.