Merged PR 35032: Avoid boxing return values in the resumed path when unnecessary
Avoid boxing return values in the resumed path when unnecessary

Place these in `Continuation.Data` instead when possible (no GC refs). It is complicated a bit by
the fact that there can already be an OSR IL offset in
`Continuation.Data`, so the return may need to be placed at offset 4;
add a flag to identify this situation. Improves `corerun.exe
varying-yields 0 1 --use-direct-yield` by about 5% according to a few
quick tests (but the variance is quite high).