Align ILLink/ILCompiler test behavior
Change most of the tests to check IsUnreferencedCodeSupported and
RequiresUnreferencedCode, with substitutions for
ILLink/ILCompiler to make them treat IsUnreferencedCodeSupported
as a constant.

There are some differences in the current ILLink/ILC test infra:

- ILC only allows embedded substitutions (not separate global
  substitutions)

- ILLink doesn't allow modifying CoreLib from test embedded
  substitutions

Because of this, we don't substitute IsDynamicCodeSupported for
ILLink (it is already substituted by default for NativeAot). This
matches the product behavior for ILLink, and leads to a small
difference (extra warning) in the tests.

We also substitute IsAssemblyFilesSupported so that NativeAot
treats it as a constant, to keep it close to the product behavior
and the analyzer behavior. For simplicity, this is done for both
NativeAot and ILLink, even though it should really be treated the
same as IsDynamicCodeCompiled for ILLink. This makes no
difference in the test behavior because we don't test how ILLink
specifically behaves with IsAssemblyFilesSupported.