[mono][interp] When local space overflows, retry compilation with optimization enabled (#94381)
The unoptimized version of a method allocates all vars outside of the execution stack as global. Some operations, like `array[,].Set`, need to reorder values on the execution stack and they do this by allocating temporary vars. These vars would be logically dead after the operation, but since we don't do any optimization they occupy the stack space for the entire method execution. Patterns where thousands of such calls are inlined in a single method by code generators can be quite common, in EFCore for example.

When the local space overflows, in addition to disabling inlining, we also try to compile with optimizations enabled.