JIT: Remove dead GT_NOP code in LSRA
Rationalization removes all GT_NOP nodes, and we do not seem to
introduce any of these before LSRA (and especially not the typed
"passthrough" GT_NOP nodes). This code is just wrong for the passthrough
nodes, so just switch all the LSRA code to only handle potential void
NOPs.

Fix #93048