Console.Unix: for echoing, write back to the terminal instead of writing to Console.Out.
On Unix .NET implements its own echo to make Console behave similar to Windows.

The echo implementation was writing to standard output, which has the unintended effect
of writing the echo characters into a redirected standard output.

This changes to write the echo to the stdin terminal where they were read from.