Merged PR 35049: Call async2 entry from thunks with direct "call" IL instruction
Call async2 entry from thunks with direct "call" IL instruction

Ensures we get proper generic lookup handling in the future when we
start supporting generics.

Currently requires that we disable inlining in the thunk due to the use
of the async continuation. It would be possible to allow inlining into
the thunk with more work.