Avoid throwing when attempting to set IInternalModelingOnlyValue (#94404)
The cctor interpreter represents some things that can only be used within the interpreter (but cannot be serialized to frozen blobs) as `IInternalModelingOnlyValue`. There was a piece of lazy code that was considering field assignments with these as invalid IL (we throw on invalid IL that we then catch later). Use the usual `Status.Fail` instead to reduce first chance exceptions. We don't want first chance exceptions in valid code.